Add fromEncoded and fromEncodedArray that accept json string
Github issue: #369

This updates the Message class to include overloads that are needed
when we have encoded messages in json format.

It also adds a check on the CypherEncode to return better error message
when the type of Data is not byte[]
