+++ src/IO.Ably.Tests.Shared/IO.Ably.Tests.Shared.projitems
@@ -109,6 +109,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Rest\StatsSandBoxSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\StatsSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Infrastructure\TestLogger.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Rest\TM3Spec.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Samples\DocumentationSamples.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Samples\RealtimeSamples.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)StatsParsingTests.cs" />
