+++ src/IO.Ably.Tests.Shared/Rest/ChannelSandboxSpecs.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 using FluentAssertions;
 using IO.Ably.Encryption;
@@ -561,123 +559,6 @@ namespace IO.Ably.Tests.Rest
             return data;
         }
 
-        [Trait("spec", "TM3")]
-        public class TM3Spec : SandboxSpecs
-        {
-            [Fact]
-            [Trait("spec", "TM3")]
-            public void Message_FromEncoded_WithNoEncoding()
-            {
-                var msg = new Message("name", "some-data");
-                var fromEncoded = Message.FromEncoded(msg);
 
-                fromEncoded.Name.Should().Be("name");
-                fromEncoded.Data.Should().Be("some-data");
-                fromEncoded.Encoding.Should().BeNullOrEmpty();
-            }
-
-            [Fact]
-            [Trait("spec", "TM3")]
-            public void Message_FromEncoded_WithEncoding()
-            {
-                var d = new Dictionary<string, string>
-                {
-                    { "foo", "bar" },
-                    { "baz", "qux" }
-                };
-
-                var msg = new Message("name", JsonHelper.Serialize(d)) { Encoding = "json" };
-                var fromEncoded = Message.FromEncoded(msg);
-
-                fromEncoded.Name.Should().Be("name");
-                JsonHelper.Serialize(fromEncoded.Data).Should().Be(JsonHelper.Serialize(d));
-                fromEncoded.Encoding.Should().BeNullOrEmpty();
-            }
-
-            [Fact]
-            [Trait("spec", "TM3")]
-            public void Message_FromEncoded_WithCustomEncoding()
-            {
-                var d = new Dictionary<string, string>
-                {
-                    { "foo", "bar" },
-                    { "baz", "qux" }
-                };
-
-                var msg = new Message("name", JsonHelper.Serialize(d)) { Encoding = "foo/json" };
-                var fromEncoded = Message.FromEncoded(msg);
-
-                fromEncoded.Name.Should().Be("name");
-                JsonHelper.Serialize(fromEncoded.Data).Should().Be(JsonHelper.Serialize(d));
-                fromEncoded.Encoding.Should().Be("foo");
-            }
-
-            [Fact]
-            [Trait("spec", "TM3")]
-            public void Message_FromEncoded_WithCipherEncoding()
-            {
-                var cipherParams = Crypto.GetDefaultParams(Crypto.GenerateRandomKey(128), null, CipherMode.CBC);
-
-                var crypto = Crypto.GetCipher(cipherParams);
-                var payload = "payload".AddRandomSuffix();
-
-                var msg = new Message("name", crypto.Encrypt(Encoding.UTF8.GetBytes(payload))) { Encoding = "utf-8/cipher+aes-128-cbc" };
-                var fromEncoded = Message.FromEncoded(msg, new ChannelOptions(cipherParams));
-
-                fromEncoded.Name.Should().Be("name");
-                fromEncoded.Data.ShouldBeEquivalentTo(payload);
-                fromEncoded.Encoding.Should().BeNullOrEmpty();
-            }
-
-            [Fact]
-            [Trait("spec", "TM3")]
-            public void Message_FromEncoded_WithInvalidCipherEncoding()
-            {
-                var cipherParams = Crypto.GetDefaultParams(Crypto.GenerateRandomKey(128), null, CipherMode.CBC);
-                var payload = "some-invalid-payload".AddRandomSuffix();
-
-                var msg = new Message("name", Encoding.UTF8.GetBytes(payload)) { Encoding = "utf-8/cipher+aes-128-cbc" };
-
-                bool didThrow = false;
-                try
-                {
-                    Message.FromEncoded(msg, new ChannelOptions(cipherParams));
-                }
-                catch (Exception)
-                {
-                    didThrow = true;
-                }
-
-                didThrow.Should().BeTrue();
-            }
-
-            [Fact]
-            [Trait("spec", "TM3")]
-            public void Message_FromEncodedArray_WithNoEncoding()
-            {
-                var msg = new[]
-                {
-                    new Message("name1", "some-data1"),
-                    new Message("name2", "some-data2")
-                };
-
-                var fromEncoded = Message.FromEncodedArray(msg);
-
-                fromEncoded.Should().HaveCount(2);
-
-                fromEncoded[0].Name.Should().Be("name1");
-                fromEncoded[0].Data.Should().Be("some-data1");
-                fromEncoded[0].Encoding.Should().BeNullOrEmpty();
-
-                fromEncoded[1].Name.Should().Be("name2");
-                fromEncoded[1].Data.Should().Be("some-data2");
-                fromEncoded[1].Encoding.Should().BeNullOrEmpty();
-            }
-
-            public TM3Spec(AblySandboxFixture fixture, ITestOutputHelper output)
-                : base(fixture, output)
-            {
-            }
-        }
     }
 }
